refactor: extract toGraph/fromGraph into utilities/stateGraph (#180)
State.ts was carrying ~440 lines of graph serialization/reconstruction
on top of the runtime state machinery. Move it to a sibling module:

- New `utilities/stateGraph.ts` holds `toGraph` and `fromGraph` plus
  their v7 callable-subtree comment block.
- `State.toGraph` and `State.fromGraph` become thin delegates to the
  extracted functions — public surface preserved, no consumer breaks.
- New `STATE_INTERNAL` Symbol on State exposes a getter/setter view
  onto private fields (`#id`, `#name`, `#bareState`,
  `#overriddenHaltState`, `#symbolToDataMap`, `#tags`) for sibling
  modules. Exported from State.ts so other files in
  `packages/machine/src` can import it; NOT re-exported from the
  package's `index.ts`, so it's off the public surface. Marked
  `@internal` in JSDoc.
- The Symbol's surface is designed with #195 already in scope:
  `collectStates` will need the same `symbolToDataMap` enumeration
  plus the existing `id` / `name` access, so the accessor doesn't
  need to grow when that lands.
- `fromGraph` uses the accessor's `name` setter to assign graph-
  sourced composite names (e.g. `A(target)`) to freshly-built bare
  States, bypassing the constructor's paren validation. That's the
  one mutation site; the setter exists for it alone and is JSDoc'd
  as such.

Import cycle State.ts ↔ stateGraph.ts resolves via ESM live bindings —
all cross-references live inside function bodies, so the bindings are
populated by the time anything reads them. Rollup flags the cycle in
the build output; it's expected.

Internal refactor, no API change. 473 tests still pass.
